#47: done #51: added ingredients delete it

+++ server/src/main/java/ua/kiev/naiv/drinkit/cocktail/web/controller/IngredientsController.java
@@ -1,7 +1,6 @@
 package ua.kiev.naiv.drinkit.cocktail.web.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,10 +10,12 @@ import org.springframework.web.bind.annotation.*;
 import ua.kiev.naiv.drinkit.cocktail.common.DrinkitUtils;
 import ua.kiev.naiv.drinkit.cocktail.common.JsonMixIn;
 import ua.kiev.naiv.drinkit.cocktail.persistence.model.Ingredient;
+import ua.kiev.naiv.drinkit.cocktail.persistence.model.IngredientWithQuantity;
 import ua.kiev.naiv.drinkit.cocktail.service.IngredientService;
 import ua.kiev.naiv.drinkit.cocktail.web.model.IngredientMixIn;
 
 import java.util.List;
+import java.util.Set;
 
 @Controller
 @RequestMapping("ingredients")
@@ -47,13 +48,19 @@ public class IngredientsController {
 
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> delete(@PathVariable int id) {
-        try {
+        Ingredient ingredient = ingredientService.getIngredientById(id);
+        if (ingredient == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Set<IngredientWithQuantity> cocktailIngredients = ingredient.getCocktailIngredients();
+        if (cocktailIngredients.isEmpty()) {
             DrinkitUtils.logOperation("Deleting ingredient", id);
             ingredientService.delete(id);
             return ResponseEntity.noContent().build();
-        } catch (EmptyResultDataAccessException e) {
-            return ResponseEntity.notFound().build();
         }
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).build();
     }
 
 }
